requirements.txt
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.3.0
celery==4.4.6
cffi==1.14.4
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==2.2.2
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-formtools==2.2
django-import-export==2.5.0
django-redis==4.12.1
django-reversion==3.0.8
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.30.2
future==0.18.2
greenlet==1.0.0
httplib2==0.9.2
importlib-metadata==3.3.0
jdcal==1.4.1
kombu==4.6.11
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.1.0
pycparser==2.20
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
pytz==2020.5
PyYAML==5.3.1
qcloudsms-py==0.1.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
vine==1.3.0
xadmin @ https://codeload.github.com/sshwsfc/xadmin/zip/django2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
